requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
python-dotenv==0.19.0
APScheduler==3.8.0
PyJWT==2.1.0
gunicorn==20.1.0
jira==3.2.0
Flask-Cors==3.0.10
openpyxl==3.0.10
pathlib==1.0.1